chore: tailwind 4, typescript 6, analytics 2, node 24, turbopack
Bumps:
- tailwindcss 3.4.19 -> 4.3.0 (with @tailwindcss/postcss)
- typescript 5.9.3 -> 6.0.3
- @vercel/analytics 1.6.1 -> 2.0.1
- @types/node 20 -> 24
- CI workflow Node 22 -> 24

Tailwind v4 migration uses @config to keep the existing v3-style
tailwind.config.js working without rewriting the Monolith palette as
CSS-first @theme. globals.css switches to @import "tailwindcss".

Drops the --webpack flag on dev and build. Turbopack now works because
MDX plugins go in as resolvable module specifiers instead of imported
functions:

- npm plugins stay as strings ("remark-gfm", "rehype-slug", etc.)
- Local plugins use absolute paths via import.meta.dirname; the
  loader's require.resolve runs with each MDX file's directory as the
  base, so relative paths break.
- remarkLLMs is a named export from fumadocs-core but the loader wants
  a default; lib/remark-llms.mjs wraps it.
- Custom annotation plugin moves from .ts to .mjs since require.resolve
  doesn't follow .ts extensions.

The @import "tw-animate-css" CSS issue from the previous PR resolves
naturally now that Tailwind v4's PostCSS plugin handles CSS resolution.
