+++ .github/workflows/ci-workflow.yml
@@ -8,7 +8,7 @@ jobs:
       - uses: pnpm/action-setup@v4
       - uses: actions/setup-node@v4
         with:
-          node-version: '22'
+          node-version: '24'
           cache: 'pnpm'
       - name: Cache Next.js build
         uses: actions/cache@v4
