+++ lib/remark-llms.mjs
@@ -1,0 +1,1 @@
+export { remarkLLMs as default } from "fumadocs-core/mdx-plugins/remark-llms";
